Changes for rename and packages
