+++ AirlineCrewManagement/src/main/java/com/airline/crewmanagement/AirlineCrewManagementApplication.java
@@ -1,9 +1,11 @@
-package com.airline.crew.management;
+package com.airline.crewmanagement;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class AirlineCrewManagementApplication {
 
 	public static void main(String[] args) {
